agentic-ai-guide/lab/requirements.txt
langchain>=0.3
langchain-core>=0.3
langchain-openai>=0.2
langchain-anthropic>=0.2
langchain-aws>=0.2
langgraph>=0.2
strands-agents>=0.1
strands-agents-tools>=0.1
mcp>=1.0
faiss-cpu>=1.8
python-dotenv>=1.0
fastapi>=0.115
uvicorn>=0.30
requests>=2.32
